Fix for event emission
Sometimes (okay well most of the time) we’d try to send an event from
Java to React before React was ready to receive the event. This adds a
check to make sure that the React Context is ready. Downside is we lose
the initial connection event (since it happens before React is loaded),
but I’ll look into fixing that in the future.